add run file for local debugging. everything should work

+++ bot.py
@@ -2,10 +2,9 @@ import asyncio
 
 import app_logger
 import json
-import os
 
 from telegram.ext import Application, MessageHandler, ContextTypes, filters, CommandHandler, CallbackContext, CallbackQueryHandler
-from telegram import Update, ForceReply, InlineKeyboardButton, InlineKeyboardMarkup, KeyboardButtonRequestUsers, helpers, User
+from telegram import Update, ForceReply, InlineKeyboardMarkup, KeyboardButtonRequestUsers, helpers, User, KeyboardButton, ReplyKeyboardMarkup
 
 from database import *
 
@@ -31,9 +30,10 @@ HELP_MESSAGE = """Komennot:
 /spin - Pyöräytä onnenpyörää
 /pullo_list - Listaa lisätyt pullot
 /pullo_add - Lisää pullo
-/pullo_remove - Poista pullo"""
+/pullo_remove - Poista pullo
+Apua: https://paihdelinkki.fi/mista-apua/"""
 
-GIF_ADD_MESSAGE = helpers.escape_markdown("Lisää gif vastaamalla tähän viestiin", 2)
+GIF_ADD_MESSAGE = helpers.escape_markdown("Lisää giffi vastaamalla tähän viestiin", 2)
 BOTTLE_ADD_MESSAGE = helpers.escape_markdown("Mikä pullo lisätään? Vastaa tähän viestiin", 2)
 ADMIN_ADD_MESSAGE = helpers.escape_markdown("Lisää admin", 2)
 
@@ -54,32 +54,31 @@ async def spin_command(update: Update, context: ContextTypes.DEFAULT_TYPE) -> No
     user = update.effective_user
     # send a random gif
     gif = gif_random()
-    message1 = await update.message.reply_text("Pyöräytetään...")
+    message1 = await update.message.reply_text("Pyöräytetään...", disable_notification=True)
     message2 = None
     if gif:
-        message2 = await update.message.reply_text(gif)
-    else:
-        message2 = await update.message.reply_text("No gifs saved.")
+        message2 = await update.message.reply_animation(gif, disable_notification=True)
     # send bot is typing action
     await update.message.reply_chat_action("typing")
     # wait 5 seconds
-    await asyncio.sleep(5)
+    await asyncio.sleep(3)
     # delete the sent gif
     await message1.delete()
-    await message2.delete()
+    if message2:
+        await message2.delete()
 
     bottle = bottle_random()
 
     mention = user.mention_markdown_v2(name=user.first_name)
 
-    response = f"Ei pulloja. Lisää komennolla /pullo_add"
+    response = helpers.escape_markdown(f"Ei pulloja. Lisää komennolla /pullo_add", 2)
     if bottle:
         if is_double:
-            response = f"Tuplashotti pullosta {bottle}!!!"
+            response = f"*{helpers.escape_markdown(f"Tuplashotti pullosta {bottle}!!!", 2)}*"
         else:
-            response = f"Shotti pullosta {bottle}!"
+            response = helpers.escape_markdown(f"Shotti pullosta {bottle}!", 2)
 
-    await update.message.reply_text(f"{mention} {helpers.escape_markdown(response, 2)}", parse_mode="MarkdownV2")
+    await update.message.reply_text(f"{mention} {response}", parse_mode="MarkdownV2")
 
 
 async def bottle_list_command(update: Update, context: ContextTypes.DEFAULT_TYPE) -> None:
@@ -88,13 +87,29 @@ async def bottle_list_command(update: Update, context: ContextTypes.DEFAULT_TYPE
 
 
 async def admin_list_command(update: Update, context: ContextTypes.DEFAULT_TYPE) -> None:
-    admins = " ".join(admin_list())
+    if not is_admin(update.effective_user.id):
+        await update.message.reply_text("Ei oikeuksia.")
+        return
+    admins = admin_list()
+    if not admins:
+        await update.message.reply_text("Ei admineja.")
+        return
+    admins = "\n".join(admin_list())
     await update.message.reply_text(admins)
 
 
 async def gif_list_command(update: Update, context: ContextTypes.DEFAULT_TYPE) -> None:
-    gifs = gif_list()
-    await update.message.reply_text(gifs)
+    if not is_admin(update.effective_user.id):
+        await update.message.reply_text("Ei oikeuksia.")
+        return
+    gif_ids, file_ids = gif_list()
+    if not gif_ids:
+        await update.message.reply_text("Ei giffejä.")
+        return
+    length = len(gif_ids)
+    for i in range(length):
+        await update.message.reply_text(f"{i + 1}")
+        await update.message.reply_animation(file_ids[i])
 
 
 user_data = {}
@@ -118,6 +133,9 @@ async def bottle_add_command(update: Update, context: ContextTypes.DEFAULT_TYPE)
 
 
 async def gif_add_command(update: Update, context: ContextTypes.DEFAULT_TYPE) -> None:
+    if not is_admin(update.effective_user.id):
+        await update.message.reply_text("Ei oikeuksia.")
+        return
     user = update.effective_user
     user_id = update.message.from_user.id
 
@@ -148,8 +166,7 @@ async def handle_user_response(update: Update, context: CallbackContext) -> None
             return
 
         # Check if the user response is empty
-        if not user_response.strip():
-            await update.message.reply_text("Yritä uudelleen.")
+        if not user_response or not user_response.strip():
             return
         bottle_name = user_response.strip()
 
@@ -174,15 +191,14 @@ async def handle_user_response(update: Update, context: CallbackContext) -> None
         if not update.message.reply_to_message or update.message.reply_to_message.message_id != user_data[user_id]["initial_message_id"]:
             return
 
-        # Check if the user response is empty
-        if not user_response.strip():
-            await update.message.reply_text("Yritä uudelleen.")
+        # Check if the user sent an animation
+        if not update.message.animation:
             return
-        gif_url = user_response.strip()
+        file_id = update.message.animation.file_id
 
         # Add the gif to your list or database here
         # Example: append to a list (replace this with your actual implementation)
-        gif_add(gif_url)
+        gif_add(file_id)
 
         # Get the initial message ID from user_data
         initial_message_id = user_data[user_id]["initial_message_id"]
@@ -191,85 +207,75 @@ async def handle_user_response(update: Update, context: CallbackContext) -> None
         await update.message.chat.delete_message(initial_message_id)
 
         # Inform the user that the gif has been added
-        await update.message.reply_text(f"Gif '{gif_url}' lisätty.")
+        await update.message.reply_text(f"Giffi lisätty.")
+
+        # Clear user state
+        del user_data[user_id]
+
+    elif state == "awaiting_admin_user":
+        # Check if the request is valid
+        if not update.message.users_shared.users:
+            return
+        admin_user_id = update.message.users_shared.users[0].user_id
+
+        # Perform the admin add operation
+        ret = admin_add(admin_user_id)
+
+        # Get the initial message ID from user_data
+        initial_message_id = user_data[user_id]["initial_message_id"]
+
+        # Delete the initial question message
+        await update.message.chat.delete_message(initial_message_id)
+
+        if not ret:
+            await update.message.reply_text(f"Admin ei lisätty.")
+        else:
+            # Inform the user that the gif has been added
+            await update.message.reply_text(f"Admin lisätty.")
 
         # Clear user state
         del user_data[user_id]
 
 
-admin_user_requests = {}
-
-
 async def admin_add_command(update: Update, context: ContextTypes.DEFAULT_TYPE) -> None:
+    if not is_admin(update.effective_user.id):
+        await update.message.reply_text("Ei oikeuksia.")
+        return
     # Create a KeyboardButtonRequestUser object
     # Create a Signed 32-bit identifier of the request
     # Store the request in the admin_user_requests dictionary
+    user_id = update.message.from_user.id
     request_id = random.randint(-2_147_483_648, 2_147_483_647)
-    button = KeyboardButtonRequestUsers(request_id, user_is_bot=False)
+    user_request = KeyboardButtonRequestUsers(request_id, user_is_bot=False)
+    button = KeyboardButton("Lisää admin", request_users=user_request)
     keyboard = [[button]]
-    reply_markup = InlineKeyboardMarkup(keyboard)
+    reply_markup = ReplyKeyboardMarkup(keyboard, one_time_keyboard=True, selective=True, resize_keyboard=True)
     message = await update.message.reply_text(ADMIN_ADD_MESSAGE, reply_markup=reply_markup, parse_mode="MarkdownV2")
     message_id = message.message_id
-    admin_user_requests[request_id] = {
-        "user_id": update.message.from_user.id,
-        "reply_markup": reply_markup,
-        "message_id": message_id
+    user_data[user_id] = {
+        "state": "awaiting_admin_user",
+        "initial_message_id": message_id,
+        "request_id": request_id,
     }
-
-
-async def handle_admin_request(update: Update, context: CallbackContext) -> None:
-    query = update.callback_query
-    user_id = query.from_user.id
-
-    # Check if the callback data exists in the admin_user_requests dictionary
-    if query.data not in admin_user_requests:
-        await query.answer("Invalid request.")
-        return
-
-    # Get the admin request
-    request = admin_user_requests[query.data]
-
-    # Check if the responding user is the one who was requested to be added as an admin
-    if user_id != request["user_id"]:
-        await query.answer("You are not authorized to respond to this request.")
-        return
-
-    # Add the user to the admin list
-    admin_add(user_id)
-
-    # Inform the user that they have been added as an admin
-    await query.message.reply_text(f"User '{user_id}' has been added as an admin.")
-
-    # Delete the original admin request message
-    await query.message.delete()
-
-    # Clear the request from the admin_user_requests dictionary
-    del admin_user_requests[query.data]
-
-    # Acknowledge the callback
-    await query.answer()
 
 
 def add_handlers():
     # Add conversation, command, and any other handlers
     # on different commands - answer in Telegram
-    application.add_handler(CommandHandler("start", start))
-    application.add_handler(CommandHandler("help", help_command))
-    application.add_handler(CommandHandler("spin", spin_command))
-    application.add_handler(CommandHandler("pullo_list", bottle_list_command))
-    application.add_handler(CommandHandler("pullo_add", bottle_add_command))
-    application.add_handler(CommandHandler("admin_list", admin_list_command))
-    application.add_handler(CommandHandler("admin_add", admin_add_command))
-    application.add_handler(CommandHandler("gif_list", gif_list_command))
-    application.add_handler(CommandHandler("gif_add", gif_add_command))
+    application.add_handler(CommandHandler("start", start, filters=~filters.UpdateType.EDITED))
+    application.add_handler(CommandHandler("help", help_command, filters=~filters.UpdateType.EDITED))
+    application.add_handler(CommandHandler("spin", spin_command, filters=~filters.UpdateType.EDITED))
+    application.add_handler(CommandHandler("pullo_list", bottle_list_command, filters=~filters.UpdateType.EDITED))
+    application.add_handler(CommandHandler("pullo_add", bottle_add_command, filters=~filters.UpdateType.EDITED))
+    application.add_handler(CommandHandler("admin_list", admin_list_command, filters=~filters.UpdateType.EDITED))
+    application.add_handler(CommandHandler("admin_add", admin_add_command, filters=~filters.UpdateType.EDITED))
+    application.add_handler(CommandHandler("gif_list", gif_list_command, filters=~filters.UpdateType.EDITED))
+    application.add_handler(CommandHandler("gif_add", gif_add_command, filters=~filters.UpdateType.EDITED))
 
     remove_commands.register_callbacks(application)
 
     # add handler for messages that are replies to our messages
-    application.add_handler(MessageHandler(filters.TEXT & ~filters.COMMAND, handle_user_response))
-
-    # add query handler for admin requests
-    application.add_handler(CallbackQueryHandler(handle_admin_request))
+    application.add_handler(MessageHandler(~filters.COMMAND & ~filters.UpdateType.EDITED, handle_user_response))
 
 
 async def run_bot(event, context):
@@ -278,3 +284,8 @@ async def run_bot(event, context):
     await application.process_update(
         Update.de_json(json.loads(event["body"]), application.bot)
     )
+
+
+def run_polling():
+    logger.info(f"Bot starting in polling mode")
+    application.run_polling()
